pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "Backend for a-to-p. Converts articles into podcasts"
authors = ["Collin Caram <collin.caram@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "api" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
pydantic-settings = "^2.1.0"
llama-index = "^0.9.28.post2"
python-dotenv = "^1.0.0"
uvicorn = "^0.25.0"
pydub = "^0.25.1"
pytest = "^7.4.4"
sqlmodel = "^0.0.14"
boto3 = "^1.34.18"
boto3-stubs = {extras = ["s3"], version = "^1.34.18"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
trafilatura = "^1.6.4"
sentry-sdk = {extras = ["fastapi"], version = "^1.39.2"}
instructor = "0.5.2"
langchain = "^0.1.3"
aiolimiter = "^1.1.0"
tenacity = "^8.2.3"
pyright = "^1.1.351"
strawberry-graphql = {extras = ["fastapi"], version = "^0.219.2"}

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.3"
pytest = "^7.4.4"
ruff = "^0.1.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
